refactor: extract selector to separate packages
